Make these percent calculations generic
Further fix the armor calculations in case the current value was the real problem.
Change the fundamental calculations in a number of ways.
Percent for a location is now an average percentage across both armor locations (if available) and the structure.
Armor percentage is 100% unless the current armor falls below 100 (make this configurable) and then it is a percentage of 100 instead of the max armor.  Well-armored mechs should resist panic until their armor is very low.
For example, a mech with 300 max armor in a location with 90 current armor would return a percentage of 90 (90/100), not 30% (90/300).
If the structure is damaged, then its percentage overrides the armor percentages unless they are lower.
So, if the front torso armor is 90% but the rear armor is gone and the structure is at 80%, then the location percentage is 80 + 0 / 2 = 40%.  It's more distressing to have your rear armor blown through and structure damage.  Less-damaged front armor wouldn't mitigate this panic.