Bump version and several fixes
Change vehicle panic to relate to the toughness of the vehicle.  Also, actually disable vehicle panic if it's disabled in the settings.
Fix panic contribution for Head.  It would panic you to have a slightly damaged head since it wasn't using (1 - value).  It was using value straight.  Not good.
Add armor threshold setting to allow mechs to ignore damage until armor falls below a certain value.  Set to zero by default in the settings and thus disabled.
Change the "minimum structure damage required" calculations to use > instead of >= because the default setting is zero for vehicles.  Taking zero structure damage would force them to make a panic save.  Not intended.